Reorganizacion de codigo
