+++ cmake-build-debug/CMakeFiles/Codigo.dir/depend.internal
@@ -5,6 +5,7 @@ CMakeFiles/Codigo.dir/estudiante.cpp.o
  /home/luisro/Documentos/Tarea-1-CPYD-s1-2021/estudiante.cpp
  /home/luisro/Documentos/Tarea-1-CPYD-s1-2021/estudiante.h
 CMakeFiles/Codigo.dir/main.cpp.o
- /home/luisro/Documentos/Tarea-1-CPYD-s1-2021/Split.h
+ /home/luisro/Documentos/Tarea-1-CPYD-s1-2021/estudiante.cpp
  /home/luisro/Documentos/Tarea-1-CPYD-s1-2021/estudiante.h
+ /home/luisro/Documentos/Tarea-1-CPYD-s1-2021/funciones.h
  /home/luisro/Documentos/Tarea-1-CPYD-s1-2021/main.cpp
