+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Apr 26 12:10 -04
+Start testing: Apr 26 13:05 -04
 ----------------------------------------------------------
-End testing: Apr 26 12:10 -04
+End testing: Apr 26 13:05 -04
