+++ main.cpp
@@ -1,69 +1,12 @@
 #include <iostream>
-#include <sstream>
-#include <fstream>
 #include <vector>
-#include <string>
 #include "estudiante.h"
-#include "Split.h"
+#include "funciones.h"
 using namespace std;
-
-
-vector<estudiante> leer_archivo(vector<estudiante> lista){
-    ifstream archivo("estudiantes.csv");
-    string linea;
-    char delimitador = ';';
-    if(archivo.is_open()){
-        cout<<"Se abrio el archivo con exito!"<<endl;
-        while(getline(archivo, linea)){
-            stringstream stream(linea);
-            string id, nombre, leng, ing, mat, cien, hist, tecn, arte, edfis;
-            int idV;
-            string nombreV;
-            float lengV, ingV, matV, cienV, histV, tecnV, arteV, edfisV;
-            getline(stream, id, delimitador);
-            getline(stream, nombre, delimitador);
-            getline(stream, leng, delimitador);
-            getline(stream, ing, delimitador);
-            getline(stream, mat, delimitador);
-            getline(stream, cien, delimitador);
-            getline(stream, hist, delimitador);
-            getline(stream, tecn, delimitador);
-            getline(stream, arte, delimitador);
-            getline(stream, edfis, delimitador);
-            // Elimino comillas dobles y se cambia tipo de var
-            idV = stoi(split(id));
-            nombreV = split(nombre);
-            lengV = stof(split(leng));
-            ingV = stof(split(ing));
-            matV = stof(split(mat));
-            cienV = stof(split(cien));
-            histV = stof(split(hist));
-            tecnV = stof(split(tecn));
-            arteV = stof(split(arte));
-            edfisV = stof(split(edfis));
-            // Crea registro estudiante, asigna valores y realiza push a lista que se debe retornar
-            estudiante reg(idV,nombreV,lengV,ingV,matV,cienV,histV,tecnV,arteV,edfisV);
-            reg.setId(idV);
-            reg.setNombre(nombreV);
-            reg.setProm_leng(lengV);
-            reg.setProm_ing(ingV);
-            reg.setProm_mat(matV);
-            reg.setProm_cien(cienV);
-            reg.setProm_hist(histV);
-            reg.setProm_tecn(tecnV);
-            reg.setProm_arte(arteV);
-            reg.setProm_edfis(edfisV);
-            lista.push_back(reg);
-        }
-        archivo.close();
-    }
-    else{cout<<"El archivo no se abrio!, debe estar en la carpeta CMake :c";}
-    return lista;
-};
 
 int main() {
     vector<estudiante> lista;
-    lista = leer_archivo(lista);
+    lista = procesar_archivo(lista);
     cout << lista[0].getNombre();
     return 0;
-};
+}
